+++ package/src/extract/path.rs
@@ -5,7 +5,7 @@ use serde::de::DeserializeOwned;
 
 use crate::response::{
     Response,
-    json::{CreateJsonResponse, JsonResponseErrorCode},
+    json::{CreateJsonResponse, JsonResponseError, JsonResponseErrorCode},
 };
 
 /// Extractor that parses path parameters.
@@ -48,8 +48,12 @@ where
             | Ok(val) => Ok(Self(val.0)),
             | Err(rej) => Err(CreateJsonResponse::failure()
                 .status(rej.status())
-                .error_code(JsonResponseErrorCode::Parse.as_str())
-                .error_message(rej.body_text())
+                .error(
+                    JsonResponseError::builder()
+                        .code(JsonResponseErrorCode::Parse.as_str())
+                        .message(rej.body_text())
+                        .build(),
+                )
                 .send()),
         }
     }
@@ -75,8 +79,12 @@ where
             | Ok(None) => Ok(None),
             | Err(rej) => Err(CreateJsonResponse::failure()
                 .status(rej.status())
-                .error_code(JsonResponseErrorCode::Parse.as_str())
-                .error_message(rej.body_text())
+                .error(
+                    JsonResponseError::builder()
+                        .code(JsonResponseErrorCode::Parse.as_str())
+                        .message(rej.body_text())
+                        .build(),
+                )
                 .send()),
         }
     }