+++ tests/src/tests/multipart/mod.rs
@@ -46,7 +46,7 @@ mod test {
 
         assert_eq!(res.success, false);
         assert_eq!(
-            res.error.unwrap().code,
+            res.errors[0].code,
             JsonResponseErrorCode::Parse.to_string()
         );
     }
@@ -63,7 +63,7 @@ mod test {
 
         assert_eq!(res.success, false);
         assert_eq!(
-            res.error.unwrap().code,
+            res.errors[0].code,
             JsonResponseErrorCode::Parse.to_string()
         );
     }
@@ -84,7 +84,7 @@ mod test {
 
         assert_eq!(res.success, false);
         assert_eq!(
-            res.error.unwrap().code,
+            res.errors[0].code,
             JsonResponseErrorCode::Parse.to_string()
         );
     }