+++ tests/src/tests/request_time_limit.rs
@@ -32,7 +32,7 @@ mod test {
 
         assert_eq!(res.success, false);
 
-        let err: JsonResponseError = res.error.unwrap();
+        let err: JsonResponseError = res.errors[0].clone();
 
         assert_eq!(err.code, JsonResponseErrorCode::Timeout.as_str());
     }